moving rails files back to previous location
