+++ test/loggerator/rails/log_subscriber_test.rb
@@ -1,4 +1,4 @@
-require_relative "../test_helper"
+require_relative "../../test_helper"
 
 class FakeEvent
   attr_accessor :payload
